+++ homeassistant/components/openhardwaremonitor/const.py
@@ -1,0 +1,3 @@
+"""Constants for the OpenHardwareMonitor integration."""
+
+DOMAIN = "openhardwaremonitor"
