--- homeassistant/components/windows_computer/const.py
@@ -1,3 +1,0 @@
-"""Constants for the Windows Computer integration."""
-
-DOMAIN = "windows_computer"
