--- homeassistant/components/windows_computer/manifest.json
@@ -1,14 +1,0 @@
-{
-  "domain": "windows_computer",
-  "name": "Windows Computer",
-  "codeowners": ["@BramKelchtermans"],
-  "config_flow": true,
-  "dependencies": [],
-  "documentation": "https://www.home-assistant.io/integrations/windows_computer",
-  "homekit": {},
-  "iot_class": "local_polling",
-  "loggers": ["custom_components.windows_computer"],
-  "requirements": [],
-  "ssdp": [],
-  "zeroconf": []
-}
