--- homeassistant/components/windows_computer/strings.json
@@ -1,20 +1,0 @@
-{
-  "config": {
-    "step": {
-      "user": {
-        "description": "Set up your OpenHardwareMonitor API to integrate with Home Assistant",
-        "data": {
-          "host": "[%key:common::config_flow::data::host%]",
-          "port": "[%key:common::config_flow::data::port%]"
-        }
-      },
-      "confirm": {
-        "description": "[%key:common::config_flow::description::confirm_setup%]"
-      }
-    },
-    "abort": {
-      "single_instance_allowed": "[%key:common::config_flow::abort::single_instance_allowed%]",
-      "no_devices_found": "[%key:common::config_flow::abort::no_devices_found%]"
-    }
-  }
-}
