--- tests/components/windows_computer/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Windows Computer integration."""
